refactor(#1464): campsite split of node database.rs into database/ submodules
Pure code motion (epic #1116): bindings/node/src/database.rs (1767L) becomes
bindings/node/src/database/{mod,options,open,write,execute_native_task,json_value}.rs,
all well under the 800-line source target. lib.rs is byte-unchanged; every
`database::…` path still resolves through mod.rs's re-exports.

The one structural change the issue asks for: the inlined single-table CQL
schema parser in the open() write-engine branch is now the named
`write::parse_single_table_schema(path: &Path)`. Everything else is a move,
plus the visibility widening the move forces (private helpers used across the
new module boundary become pub(super)) and per-module import splits.

The deprecated `value_to_json` is isolated in json_value.rs so #1457's
next-major delete is a single-file removal; it is NOT deleted here.
